+++ spec/models/post_spec.rb
@@ -1,5 +1,7 @@
 require 'rails_helper'
 
+include TestFactories
+
 describe Post do 
   describe "vote methods" do
 
@@ -36,25 +38,5 @@ describe Post do
       expect( post.up_votes ).to eq(1)
     end
   end
-
-
- def associated_post(options={})
-   post_options = {
-    title: 'Post title',
-    body: 'Post bodies must be pretty long.',
-    topic: Topic.create(name: 'Topic name'),
-    user: authenticated_user
-    }.merge(options)
-
-    Post.create(post_options)
- end
-
-  def authenticated_user(options={})
-     user_options = {email: "email#{rand}@fake.com", password: 'password'}.merge(options)
-     user = User.new(user_options)
-     user.skip_confirmation!
-     user.save
-     user
-   end
 
 end
