+++ spec/models/vote_spec.rb
@@ -1,5 +1,7 @@
 require 'rails_helper'
 
+include TestFactories
+
 describe Vote do
   describe "validations" do
     describe "value validation" do
@@ -25,25 +27,5 @@ describe Vote do
        vote.save
      end
    end
-
-   def associated_post(options={})
-   post_options = {
-    title: 'Post title',
-    body: 'Post bodies must be pretty long.',
-    topic: Topic.create(name: 'Topic name'),
-    user: authenticated_user
-    }.merge(options)
-
-    Post.create(post_options)
- end
-
-  def authenticated_user(options={})
-     user_options = {email: "email#{rand}@fake.com", password: 'password'}.merge(options)
-     user = User.new(user_options)
-     user.skip_confirmation!
-     user.save
-     user
-   end
-
 
 end
